setup posts and remove old about

+++ db/migrate/20140705015425_drop_posts.rb
@@ -1,5 +1,5 @@
 class DropPosts < ActiveRecord::Migration
   def change
-    drop_table :posts
+    drop_table :posts if table_exists? :posts
   end
 end
